requirements.txt
aiohttp==3.8.6
asyncpg==0.28.0
openpyxl==3.1.2
pydantic==2.4.2
pydantic-sqlalchemy==0.0.9
SQLAlchemy==2.0.22
python-decouple==3.8
greenlet==3.0.0
google-cloud-storage==2.12.0